Released V 0.17 Fix to avoid empty error messages.
git-svn-id: svn://projects.linuxtogo.org/svn/gpe/trunk@6162 d4735bf7-8e1f-0410-bce4-ef584eb11c25

+++ base/gpe-su/gpe-su.c
@@ -113,7 +113,8 @@ do_su (void)
   *rp = 0;
 
   if ((strncmp (result, "Password:", 9))
-	  && (!strstr(result, "Gdk-WARNING")))
+	  && (!strstr(result, "Gdk-WARNING"))
+ 	  && (strlen(g_strstrip(result))))
     {
       gpe_error_box (result);
       kill (pid, 15);
